+++ examples/deno/deps.ts
@@ -1,12 +1,12 @@
 export {
   composeStrategy,
   createDateStrategy,
-  fromJsonAs,
-  FromJsonStrategy,
+  fromJSONAs,
+  FromJSONStrategy,
   iso8601Date,
   Serializable,
   SerializeProperty,
-  ToJsonStrategy,
+  ToJSONStrategy,
   TransformKey,
 } from "https://deno.land/x/ts_serialize/mod.ts";
-export { readJson } from "https://deno.land/std@0.68.0/fs/mod.ts";
+export { readJSON } from "https://deno.land/std@0.68.0/fs/mod.ts";