+++ from_json/default.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
-import { JsonValue } from "../serializable.ts";
+import { JSONValue } from "../serializable.ts";
 
 /** typed function */
-export function defaultFromJson(value: JsonValue): any {
+export function defaultFromJSON(value: JSONValue): any {
   return value;
 }