+++ from_json/default_test.ts
@@ -1,11 +1,11 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
 import { assertEquals, test } from "../test_deps.ts";
-import { defaultFromJson } from "./default.ts";
+import { defaultFromJSON } from "./default.ts";
 
 test({
-  name: "defaultFromJson passes the value unmodified",
+  name: "defaultFromJSON passes the value unmodified",
   fn() {
-    assertEquals(defaultFromJson("Hello world"), "Hello world");
+    assertEquals(defaultFromJSON("Hello world"), "Hello world");
   },
 });