+++ to_json/default.ts
@@ -1,10 +1,10 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
-import { JsonValue } from "../serializable.ts";
+import { JSONValue } from "../serializable.ts";
 
 /** Use the default replacer logic 
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#The_replacer_parameter
 */
-export function defaultToJson(value: any): JsonValue {
+export function defaultToJSON(value: any): JSONValue {
   return value;
 }