+++ to_json/recursive.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
-import { JsonObject, Serializable, toPojo } from "../serializable.ts";
+import { JSONObject, Serializable, toPojo } from "../serializable.ts";
 
 /** Recursively serialize a serializable class */
-export function recursiveToJson(value: Serializable): JsonObject {
+export function recursiveToJSON(value: Serializable): JSONObject {
   return toPojo(value);
 }